generalize nspv daemon start
